+++ github3/repos/repo.py
@@ -1307,8 +1307,14 @@ class Repository(GitHubCore):
                     etag=None):
         """Iterate over issues on this repo based upon parameters passed.
 
+        .. versionchanged:: 0.9.0
+
+            The ``state`` parameter now accepts 'all' in addition to 'open'
+            and 'closed'.
+
         :param int milestone: (optional), 'none', or '*'
-        :param str state: (optional), accepted values: ('open', 'closed')
+        :param str state: (optional), accepted values: ('all', 'open',
+            'closed')
         :param str assignee: (optional), 'none', '*', or login name
         :param str mentioned: (optional), user's login name
         :param str labels: (optional), comma-separated list of labels, e.g.
@@ -1459,15 +1465,29 @@ class Repository(GitHubCore):
                 del params[k]
         return self._iter(int(number), url, Thread, params, etag)
 
-    def iter_pulls(self, state=None, head=None, base=None, number=-1,
-                   etag=None):
+    def iter_pulls(self, state=None, head=None, base=None, sort='created',
+                   direction='desc', number=-1, etag=None):
         """List pull requests on repository.
 
-        :param str state: (optional), accepted values: ('open', 'closed')
+        .. versionchanged:: 0.9.0
+
+            - The ``state`` parameter now accepts 'all' in addition to 'open'
+              and 'closed'.
+
+            - The ``sort`` parameter was added.
+
+            - The ``direction`` parameter was added.
+
+        :param str state: (optional), accepted values: ('all', 'open',
+            'closed')
         :param str head: (optional), filters pulls by head user and branch
             name in the format ``user:ref-name``, e.g., ``seveas:debian``
         :param str base: (optional), filter pulls by base branch name.
             Example: ``develop``.
+        :param str sort: (optional), Sort pull requests by ``created``,
+            ``updated``, ``popularity``, ``long-running``. Default: 'created'
+        :param str direction: (optional), Choose the direction to list pull
+            requests. Accepted values: ('desc', 'asc'). Default: 'desc'
         :param int number: (optional), number of pulls to return. Default: -1
             returns all available pull requests
         :param str etag: (optional), ETag from a previous request to the same
@@ -1477,9 +1497,9 @@ class Repository(GitHubCore):
         """
         url = self._build_url('pulls', base_url=self._api)
         params = {}
-        if state and state.lower() in ('open', 'closed'):
+        if state and state.lower() in ('all', 'open', 'closed'):
             params['state'] = state.lower()
-        params.update(head=head, base=base)
+        params.update(head=head, base=base, sort=sort, direction=direction)
         self._remove_none(params)
         return self._iter(int(number), url, PullRequest, params, etag)