+++ tests/integration/test_repos_repo.py
@@ -38,6 +38,15 @@ class TestRepository(IntegrationHelper):
             for d in repository.iter_deployments():
                 assert isinstance(d, github3.repos.deployment.Deployment)
 
+    def test_iter_issues_accepts_state_all(self):
+        """Test that the state parameter accets 'all'."""
+        cassette_name = self.cassette_name('issues_state_all')
+        with self.recorder.use_cassette(cassette_name):
+            repository = self.gh.repository('sigmavirus24', 'betamax')
+            assert repository is not None
+            for issue in repository.iter_issues(state='all'):
+                assert issue.state in ('open', 'closed')
+
     def test_iter_languages(self):
         """Test that a repository's languages can be retrieved."""
         cassette_name = self.cassette_name('iter_languages')
@@ -48,6 +57,19 @@ class TestRepository(IntegrationHelper):
                 assert 'ETag' not in l
                 assert 'Last-Modified' not in l
                 assert isinstance(l, tuple)
+
+    def test_iter_pulls_accepts_sort_and_direction(self):
+        """Test that iter_pulls now takes a sort parameter."""
+        cassette_name = self.cassette_name('pull_requests_accept_sort')
+        with self.recorder.use_cassette(cassette_name):
+            repository = self.gh.repository('sigmavirus24', 'betamax')
+            assert repository is not None
+            last_pr = None
+            for pr in repository.iter_pulls(sort='updated', direction='asc'):
+                assert pr
+                if last_pr:
+                    assert last_pr.updated_at < pr.updated_at
+                last_pr = pr
 
     def test_iter_releases(self):
         """Test the ability to iterate over releases on a repository."""