+++ HISTORY.rst
@@ -1,6 +1,27 @@
 History/Changelog
 -----------------
 
+0.9.0: 2014-03-xx
+~~~~~~~~~~~~~~~~~
+
+- Add Deployments API
+
+- Add support so applications can revoke a `single authorization`_ or `all
+  authorizations`_ created by the application
+
+- Add the ability for users to ping_ hooks
+
+- Allow users to list a `Repository's collaborators`_
+
+- Update how users can list issues and pull requests. See:
+  http://developer.github.com/changes/2014-02-28-issue-and-pull-query-enhancements/
+  This includes breaking changes to ``Repository#iter_pulls``.
+
+.. _single authorization: https://github3py.readthedocs.org/en/latest/github.html#github3.github.GitHub.revoke_authorization
+.. _all authorizations: https://github3py.readthedocs.org/en/latest/github.html#github3.github.GitHub.revoke_authorizations
+.. _ping: https://github3py.readthedocs.org/en/latest/repos.html?highlight=ping#github3.repos.hook.Hook.ping
+.. _Repository's collaborators: https://github3py.readthedocs.org/en/latest/repos.html#github3.repos.repo.Repository.iter_collaborators
+
 0.8.2: 2014-02-11
 ~~~~~~~~~~~~~~~~~